+++ packages/utils/src/handleInterceptorCode/HandleInterceptorCodeCom.tsx
@@ -1,16 +1,16 @@
 /*
  * @Author: hbwow lllengkaixin@gmail.com
  * @Date: 2024-01-24 15:35:03
- * @LastEditors: hbwow lllengkaixin@gmail.com
- * @LastEditTime: 2024-01-24 16:50:05
+ * @LastEditors: 冷开俊 lengkj@travelsky.com.cn
+ * @LastEditTime: 2024-10-16 11:09:55
  * @FilePath: /more-repo/packages/utils/src/handleInterceptorCode/handleInterceptorCodeCom.tsx
  * @Description: 因为直接静态方法使用Modal会导致拿不到context上下文，所以写个组件搭配使用
  */
-import { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import { Modal, ModalFuncProps } from 'antd';
 
-import { events } from '../utils/eventEmit';
+import { emitter } from '../utils/eventEmit';
 import { EVENTS_NAME } from './handleInterceptorCode';
 
 interface IModalProps extends ModalFuncProps {
@@ -18,30 +18,44 @@ interface IModalProps extends ModalFuncProps {
   modalDestroyAll?: boolean; // 销毁所有弹窗
 }
 
-const HandleInterceptorCodeCom = () => {
+export interface IHandleInterceptorCodeComProps {
+  customRender?: (props: any) => React.ReactNode; // 自定义渲染
+}
+
+const HandleInterceptorCodeCom = ({ customRender }: IHandleInterceptorCodeComProps) => {
   const [modal, contextHolder] = Modal.useModal();
+  const [modalProps, setModalProps] = useState<IModalProps>();
 
   useEffect(() => {
-    events.on(
-      EVENTS_NAME,
-      ({ modalMethod = 'warning', modalDestroyAll, onOk, ...rest }: IModalProps) => {
-        const nextProps: IModalProps = { ...rest };
+    const listener = (_modalProps: IModalProps) => {
+      if (customRender) {
+        setModalProps(_modalProps);
+      }
 
-        if (onOk) {
-          nextProps.onOk = () => {
-            onOk();
-            if (modalDestroyAll) {
-              Modal.destroyAll();
-            }
-          };
-        }
+      const { modalMethod = 'warning', modalDestroyAll, onOk, ...rest } = _modalProps;
 
-        modal[modalMethod]({ ...nextProps });
-      },
-    );
-  }, [modal]);
+      const nextProps: IModalProps = { ...rest };
 
-  return <>{contextHolder}</>;
+      if (onOk) {
+        nextProps.onOk = () => {
+          onOk();
+          if (modalDestroyAll) {
+            Modal.destroyAll();
+          }
+        };
+      }
+
+      modal[modalMethod]({ ...nextProps });
+    };
+
+    emitter.on(EVENTS_NAME, listener);
+
+    return () => {
+      emitter.off(EVENTS_NAME, listener);
+    };
+  }, [modal, customRender]);
+
+  return <>{customRender ? customRender({ ...modalProps }) : contextHolder}</>;
 };
 
 export default HandleInterceptorCodeCom;