+++ packages/components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hbwow/components
 
+## 2.18.0
+
+### Minor Changes
+
+- cloneChildren
+
 ## 2.17.0
 
 ### Minor Changes